app/requirements.txt
requests==2.32.3
mysql-connector-python==9.0.0

